+++ packages/core/src/connectors/injected.ts
@@ -34,8 +34,6 @@ export type InjectedConnectorOptions = {
   shimDisconnect?: boolean
 }
 
-export const shimDisconnectKey = 'injected.shimDisconnect'
-
 export class InjectedConnector extends Connector<
   Window['ethereum'],
   InjectedConnectorOptions | undefined,
@@ -48,6 +46,8 @@ export class InjectedConnector extends Connector<
   #provider?: Window['ethereum']
   #switchingChains?: boolean
 
+  protected shimDisconnectKey = 'injected.shimDisconnect'
+
   constructor({
     chains,
     options = { shimDisconnect: true },
@@ -58,20 +58,18 @@ export class InjectedConnector extends Connector<
     super({ chains, options })
 
     let name = 'Injected'
-    if (typeof window !== 'undefined') {
-      const overrideName = options.name
+    const overrideName = options.name
+    if (typeof overrideName === 'string') name = overrideName
+    else if (typeof window !== 'undefined') {
       const detectedName = getInjectedName(window.ethereum)
-      if (overrideName)
-        name =
-          typeof overrideName === 'function'
-            ? overrideName(detectedName)
-            : overrideName
+      if (overrideName) name = overrideName(detectedName)
       else
         name =
           typeof detectedName === 'string'
             ? detectedName
             : <string>detectedName[0]
     }
+
     this.id = 'injected'
     this.name = name
   }
@@ -101,7 +99,7 @@ export class InjectedConnector extends Connector<
 
       // Add shim to storage signalling wallet is connected
       if (this.options?.shimDisconnect)
-        getClient().storage?.setItem(shimDisconnectKey, true)
+        getClient().storage?.setItem(this.shimDisconnectKey, true)
 
       return { account, chain: { id, unsupported }, provider }
     } catch (error) {
@@ -123,7 +121,7 @@ export class InjectedConnector extends Connector<
 
     // Remove shim signalling wallet is disconnected
     if (this.options?.shimDisconnect)
-      getClient().storage?.removeItem(shimDisconnectKey)
+      getClient().storage?.removeItem(this.shimDisconnectKey)
   }
 
   async getAccount() {
@@ -165,7 +163,7 @@ export class InjectedConnector extends Connector<
       if (
         this.options?.shimDisconnect &&
         // If shim does not exist in storage, wallet is disconnected
-        !getClient().storage?.getItem(shimDisconnectKey)
+        !getClient().storage?.getItem(this.shimDisconnectKey)
       )
         return false
 
@@ -290,7 +288,7 @@ export class InjectedConnector extends Connector<
     this.emit('disconnect')
     // Remove shim signalling wallet is disconnected
     if (this.options?.shimDisconnect)
-      getClient().storage?.removeItem(shimDisconnectKey)
+      getClient().storage?.removeItem(this.shimDisconnectKey)
   }
 
   protected isUserRejectedRequestError(error: unknown) {