+++ packages/react/src/hooks/contracts/useContractRead.ts
@@ -3,12 +3,12 @@ import {
   ReadContractConfig,
   ReadContractResult,
   deepEqual,
+  parseContractResult,
   readContract,
 } from '@wagmi/core'
 import { hashQueryKey } from 'react-query'
 
 import { QueryConfig, QueryFunctionArgs } from '../../types'
-import { parseContractResult } from '../../utils'
 import { useBlockNumber } from '../network-status'
 import { useChainId, useInvalidateOnBlock, useQuery } from '../utils'
 
@@ -44,7 +44,7 @@ const queryKeyHashFn = ([queryKey_]: ReturnType<typeof queryKey>) => {
   return hashQueryKey([rest])
 }
 
-const queryFn = ({
+const queryFn = async ({
   queryKey: [
     {
       addressOrName,
@@ -56,14 +56,16 @@ const queryFn = ({
     },
   ],
 }: QueryFunctionArgs<typeof queryKey>) => {
-  return readContract({
-    addressOrName,
-    args,
-    chainId,
-    contractInterface,
-    functionName,
-    overrides,
-  })
+  return (
+    (await readContract({
+      addressOrName,
+      args,
+      chainId,
+      contractInterface,
+      functionName,
+      overrides,
+    })) || null
+  )
 }
 
 export function useContractRead({