+++ packages/react/src/hooks/providers/useWebSocketProvider.test.ts
@@ -30,5 +30,25 @@ describe('useWebSocketProvider', () => {
         `"<WebSocketProvider network={1} />"`,
       )
     })
+
+    it('switches chainId', async () => {
+      const client = setupClient({
+        webSocketProvider: getWebSocketProvider,
+      })
+      await client.webSocketProvider?.destroy()
+      let chainId = 1
+      const { result } = renderHook(() => useWebSocketProvider({ chainId }), {
+        initialProps: { client },
+      })
+      await result.current?.destroy()
+      expect(result.current).toMatchInlineSnapshot(
+        `"<WebSocketProvider network={1} />"`,
+      )
+      chainId = 4
+      await result.current?.destroy()
+      expect(result.current).toMatchInlineSnapshot(
+        `"<WebSocketProvider network={1} />"`,
+      )
+    })
   })
 })